Organize queryString and string util exports
