+++ src/hooks/tests/useUpdateUrlSortParamOnTableChange/TestComponent.tsx
@@ -3,8 +3,9 @@ import { Table } from "antd";
 import { ColumnProps } from "antd/es/table";
 import { useLocation } from "react-router";
 import { useUpdateUrlSortParamOnTableChange } from "hooks";
-import { parseQueryString } from "utils";
+import { queryString } from "utils";
 
+const { parseQueryString } = queryString;
 export const TestComponent = () => {
   const tableChangeHandler = useUpdateUrlSortParamOnTableChange();
