+++ src/pages/Hosts.tsx
@@ -25,8 +25,10 @@ import {
 } from "gql/generated/types";
 import { HOSTS } from "gql/queries";
 import { HostsTable } from "pages/hosts/HostsTable";
-import { parseQueryString, getArray, getString } from "utils";
+import { queryString } from "utils";
 import { getPageFromSearch, getLimitFromSearch } from "utils/url";
+
+const { getArray, getString, parseQueryString } = queryString;
 
 export const Hosts: React.FC = () => {
   const hostsTableAnalytics = useHostsTableAnalytics();
