+++ src/pages/spawn/spawnHost/SpawnHostTable.tsx
@@ -8,10 +8,12 @@ import { HostStatusBadge } from "components/HostStatusBadge";
 import { DoesNotExpire, SpawnTable } from "components/Spawn";
 import { WordBreak } from "components/Typography";
 import { MyHost } from "types/spawn";
-import { parseQueryString } from "utils";
+import { queryString } from "utils";
 import { sortFunctionDate, sortFunctionString } from "utils/string";
 import { SpawnHostCard } from "./SpawnHostCard";
 import { SpawnHostTableActions } from "./SpawnHostTableActions";
+
+const { parseQueryString } = queryString;
 
 interface SpawnHostTableProps {
   hosts: MyHost[];
