+++ src/utils/index.ts
@@ -1,8 +1,4 @@
-export {
-  parseQueryString,
-  getArray,
-  getString,
-  queryParamAsNumber,
-} from "utils/queryString/parseQueryString";
-export { stringifyQuery } from "utils/queryString/stringifyQuery";
-export { isFailedTaskStatus } from "utils/statuses/isFailedTaskStatus";
+import * as statuses from "utils/statuses";
+import * as queryString from "./queryString";
+
+export { queryString, statuses };
