+++ src/utils/statuses/index.ts
@@ -1,0 +1,4 @@
+import { getStatusBadgeCopy } from "./getStatusBadgeCopy";
+import { isFailedTaskStatus } from "./isFailedTaskStatus";
+
+export { isFailedTaskStatus, getStatusBadgeCopy };
