+++ src/utils/string/index.ts
@@ -164,5 +164,3 @@ export const sortFunctionDate = (a, b, key) => {
   }
   return dateA - dateB;
 };
-
-export { getStatusBadgeCopy } from "./getStatusBadgeCopy";
